+++ flowy/local/runner.py
@@ -1,26 +1,10 @@
-try:
-    from concurrent.futures import ProcessPoolExecutor
-except ImportError:
-    from futures import ProcessPoolExecutor
-
 import copy
 import json
-import warnings
-from collections import namedtuple
 from functools import partial
 from threading import Event
 from threading import RLock
 
-from flowy.swf.history import SWFTaskExecutionHistory as TaskHistory
-from flowy.serialization import JSONProxyEncoder
-from flowy.proxy import Proxy
-from flowy.tracer import ExecutionTracer
 from flowy.result import TaskError
-from flowy.tracer import TracingProxy
-from flowy.worker import Worker
-from flowy.config import WorkflowConfig
-
-__all__ = ['LocalWorkflow']
 
 
 class WorkflowRunner(object):
@@ -334,149 +318,3 @@ class State(object):
             len(self.running), len(self.results), len(self.errors), order)
 
 
-class Decision(dict):
-    def __init__(self):
-        self['type'] = 'schedule'
-        self['activities'] = []
-        self['workflows'] = []
-        self.closed = False
-
-    def fail(self, reason):
-        if self.closed:
-            return
-        self.clear()
-        self['type'] = 'fail'
-        self['reason'] = reason
-        self.closed = True
-
-    def flush(self):
-        self.closed = True
-
-    def restart(self, input_data):
-        if self.closed:
-            return
-        self.clear()
-        self['type'] = 'restart'
-        self['input_data'] = input_data
-        self.closed = True
-
-    def finish(self, result):
-        if self.closed:
-            return
-        self.clear()
-        self['type'] = 'finish'
-        self['result'] = result
-        self.closed = True
-
-    def schedule_activity(self, call_key, input_data, f):
-        if self.closed or 'activities' not in self:
-            return
-        self['activities'].append(
-            {'id': call_key,
-             'input_data': input_data,
-             'f': f})
-
-    def schedule_workflow(self, call_key, input_data, f):
-        if self.closed or 'workflows' not in self:
-            return
-        self['workflows'].append(
-            {'id': call_key,
-             'input_data': input_data,
-             'f': f})
-
-
-class ActivityDecision(object):
-    def __init__(self, decision, identity, f):
-        self.decision = decision
-        self.identity = identity
-        self.f = f
-
-    def fail(self, reason):
-        self.decision.fail(reason)
-
-    def schedule(self, call_number, retry_number, delay, input_data):
-        self.decision.schedule_activity(
-            '%s-%s-%s' % (self.identity, call_number, retry_number),
-            input_data, self.f)
-
-
-class WorkflowDecision(object):
-    def __init__(self, decision, identity, f):
-        self.decision = decision
-        self.identity = identity
-        self.f = f
-
-    def fail(self, reason):
-        self.decision.fail(reason)
-
-    def schedule(self, call_number, retry_number, delay, input_data):
-        self.decision.schedule_workflow(
-            '%s-%s-%s' % (self.identity, call_number, retry_number),
-            input_data, self.f)
-
-
-class ActivityProxy(object):
-    def __init__(self, identity, f):
-        self.identity = identity
-        self.f = f
-
-    def __call__(self, decision, history, tracer):
-        th = TaskHistory(history, self.identity)
-        ad = ActivityDecision(decision, self.identity, self.f)
-        if tracer is None:
-            return Proxy(th, ad)
-        return TracingProxy(tracer, self.identity, th, ad)
-
-
-class WorkflowProxy(object):
-    def __init__(self, identity, f):
-        self.identity = identity
-        self.f = f
-
-    def __call__(self, decision, history, tracer):
-        th = TaskHistory(history, self.identity)
-        wd = WorkflowDecision(decision, self.identity, self.f)
-        if tracer is None:
-            return Proxy(th, wd)
-        return TracingProxy(tracer, self.identity, th, wd)
-
-
-class LocalWorkflow(WorkflowConfig):
-    def __init__(self, w,
-                 activity_workers=8,
-                 workflow_workers=2,
-                 executor=ProcessPoolExecutor):
-        super(LocalWorkflow, self).__init__()
-        self.activity_workers = activity_workers
-        self.workflow_workers = workflow_workers
-        self.executor = executor
-        self.worker = Worker()
-        self.worker.register_task('local', self.wrap(w))
-
-    def conf_activity(self, dep_name, f):
-        self.conf_proxy_factory(dep_name, ActivityProxy(dep_name, f))
-
-    def conf_workflow(self, dep_name, f):
-        self.conf_proxy_factory(dep_name, WorkflowProxy(dep_name, f))
-
-    def __call__(self, state, input_data, tracer):
-        # NB: The final trace can be computed only on the last decision
-        # thread/process
-        d = Decision()
-        self.worker('local', input_data, d,
-                    d, state, tracer) # pass to proxies
-        if d['type'] in ['finish', 'fail'] and tracer is not None:
-            tracer.display()
-        return d
-
-    def run(self, *args, **kwargs):
-        wait = kwargs.pop('_wait', False)
-        tracer = None
-        if kwargs.pop('_trace', False):
-            tracer = ExecutionTracer()
-        a_executor = self.executor(max_workers=self.activity_workers)
-        w_executor = self.executor(max_workers=self.workflow_workers)
-        input_data = Proxy.serialize_input(*args, **kwargs)
-        wr = RootWorkflowRunner(self, w_executor, a_executor, input_data,
-                                tracer=tracer)
-        return wr.run(wait=wait)
